requirements.txt
streamlit
redis
pandas
plotly
paramiko
groq
kafka-python
python-dotenv